# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/spaces": {
      "get": {
        "operationId": "getSpacesTree",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SpaceDto"
                  }
                }
              }
            }
          }
        },
        "summary": "",
        "tags": [
          "spaces"
        ]
      },
      "post": {
        "operationId": "createSpace",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSpaceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpaceDetailDto"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": [
          "spaces"
        ]
      }
    },
    "/spaces/{id}": {
      "get": {
        "operationId": "getSpace",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpaceDetailDto"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": [
          "spaces"
        ]
      },
      "patch": {
        "operationId": "updateSpace",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSpaceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpaceDetailDto"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": [
          "spaces"
        ]
      },
      "delete": {
        "operationId": "deleteSpace",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "",
        "tags": [
          "spaces"
        ]
      }
    },
    "/spaces/{id}/members": {
      "get": {
        "operationId": "getSpaceMembers",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SpaceMemberDto"
                  }
                }
              }
            }
          }
        },
        "summary": "",
        "tags": [
          "spaces"
        ]
      }
    },
    "/spaces/{id}/inheritance": {
      "put": {
        "operationId": "updateSpaceInheritance",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSpaceInheritanceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpaceDetailDto"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": [
          "spaces"
        ]
      }
    },
    "/spaces/{id}/members/{userId}": {
      "put": {
        "operationId": "addSpaceMember",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "",
        "tags": [
          "spaces"
        ]
      },
      "delete": {
        "operationId": "removeSpaceMember",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "",
        "tags": [
          "spaces"
        ]
      }
    },
...
  },
  "info": {
    "title": "Folioteca API",
    "description": "Contrato público da API. Gerado por @nestjs/swagger.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "SpaceDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "kind": {
            "type": "string",
            "enum": [
              "UNIT",
              "FREE"
            ]
          },
          "name": {
            "type": "string"
          },
          "unitId": {
            "type": "string",
            "nullable": true
          },
          "restricted": {
            "type": "boolean"
          },
          "inheritsFromParent": {
            "type": "boolean"
          },
          "managerId": {
            "type": "string",
            "nullable": true
          },
          "children": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SpaceDto"
            }
          }
        },
        "required": [
          "id",
          "kind",
          "name",
          "unitId",
          "restricted",
          "inheritsFromParent",
          "managerId",
          "children"
        ]
      },
      "SpacePathEntryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "SpaceDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "kind": {
            "type": "string",
            "enum": [
              "UNIT",
              "FREE"
            ]
          },
          "name": {
            "type": "string"
          },
          "unitId": {
            "type": "string",
            "nullable": true
          },
          "parentId": {
            "type": "string",
            "nullable": true
          },
          "restricted": {
            "type": "boolean"
          },
          "inheritsFromParent": {
            "type": "boolean"
          },
          "managerId": {
            "type": "string",
            "nullable": true
          },
          "path": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SpacePathEntryDto"
            }
          }
        },
        "required": [
          "id",
          "kind",
          "name",
          "unitId",
          "parentId",
          "restricted",
          "inheritsFromParent",
          "managerId",
          "path"
        ]
      },
      "SpaceMemberDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "image": {
            "type": "string",
            "nullable": true
          },
          "isManager": {
            "type": "boolean"
          }
        },
        "required": [
          "userId",
          "name",
          "email",
          "image",
          "isManager"
        ]
      },
      "CreateSpaceDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Nome do espaço livre."
          },
          "parentId": {
            "type": "string",
            "nullable": true,
            "description": "Id do espaço pai (unidade ou livre), ou null para criar no topo."
          },
          "restricted": {
            "type": "boolean",
            "description": "Restrito — só quem for convidado vê este espaço."
          }
        },
        "required": [
          "name",
          "parentId"
        ]
      },
      "UpdateSpaceDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Novo nome do espaço."
          },
          "restricted": {
            "type": "boolean",
            "description": "Restrito — só quem for convidado vê este espaço."
          }
        }
      },
      "UpdateSpaceInheritanceDto": {
        "type": "object",
        "properties": {
          "inheritsFromParent": {
            "type": "boolean",
            "description": "Liga ou desliga a herança do compartilhamento do espaço acima (M11)."
          }
        },
        "required": [
          "inheritsFromParent"
        ]
      },
...
    }
  }
}